final_version_1.1

+++ controllers/programController.js
@@ -415,15 +415,17 @@ class ProgramController {
         if (!themes) {
             return next(ApiError.internal( 'Темы для удаления не найдены'))
         }
-      
+        try {
         themes.forEach(async (theme) => {
             if (theme.presetation_src) {
-                    
-                fs.unlink(path.resolve(__dirname, '..', 'static', theme.presetation_src), (err) => {
-                    if (err) throw err;
-
-                    
-                }); 
+                if (fs.existsSync(__dirname, '..', 'static', theme.presetation_src)) {
+                    fs.unlink(path.resolve(__dirname, '..', 'static', theme.presetation_src), (err) => {
+                        if (err) throw err;
+    
+                        
+                    }); 
+                }
+                
             }
             let puncts = await Punct.findAll({where: {themeId: theme.id}})
             if (!puncts) {
@@ -432,12 +434,13 @@ class ProgramController {
             puncts.forEach(async (punct) => {
                 try {
                     if (punct.lection_src) {
-                    
-                        fs.unlink(path.resolve(__dirname, '..', 'static', punct.lection_src), (err) => {
-                            if (err) throw err;
-    
-                          
-                        }); 
+                        if (fs.existsSync(__dirname, '..', 'static', theme.lection_src)) {
+                            fs.unlink(path.resolve(__dirname, '..', 'static', punct.lection_src), (err) => {
+                                if (err) throw err;
+        
+                            
+                            }); 
+                        }
                     } 
                 } catch(e) {
                     return next(ApiError.internal( 'Ошибка при удалении файлов с сервера'))
@@ -450,6 +453,9 @@ class ProgramController {
          
             
         })
+        } catch(e) {
+            return next(ApiError.badRequest( 'Файлы для удаления не были найдены'))
+        }
         await Program.destroy({
             where: {
               id: id,
